Fixed detekt.
